requirements.txt
python-dotenv
requests
Flask
